# API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: SIAR Platform ERP/OMS/WMS API
  description: Enterprise-grade Order Management, Inventory Control, and Fulfillment API
  version: 1.0.0
  contact:
    name: SIAR Digital Support
    email: support@siardigital.com
  license:
    name: Proprietary
    url: https://www.siardigital.com
servers:
  - url: http://localhost:5000/api
    description: Development Server
  - url: https://erp.siardigital.com/api
    description: Production Server
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create new order
      description: Create an order from any marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: integer
                marketplace:
                  type: string
                  enum: [Shopify, Amazon, TikTok, Walmart, eBay, Etsy]
                marketplace_order_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                total_amount:
                  type: number
      responses:
        '201':
          description: Order created
        '400':
          description: Invalid order data

    get:
      tags:
        - Orders
      summary: List orders
      description: Get orders with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: marketplace
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Orders list
...
security:
  - BearerToken: []

# =====================================================================
# COMPONENTS & SCHEMAS
# =====================================================================
...